requirements.txt
Django==2.2.6
django-active-link==0.1.6
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.7
django-colorfield==0.1.15
django-gravatar2==1.4.2
django-http2-push==0.0b2
django-markdown-deux==1.0.5
django-markdown-filter==0.0.1
django-pipeline==1.6.14
django-progressive-web-app==0.1.1
django-pwa==1.0.5
django-recaptcha3==0.1.0
django-scheduler==0.8.8
django-webpush==0.3.1
djangorestframework==3.9.0
gunicorn==19.9.0
markdown2==2.3.7
psycopg2==2.7.7
psycopg2-binary==2.8.3
pywebpush==1.9.4
requests==2.22.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
six==1.10.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
tika==1.19
urllib3==1.25.6
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
idna==2.8
oauthlib==3.1.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
soupsieve==1.9.4
